Create ClientScopesTable and RoleScopesTable components
Each of these components has the abiltiy to render
a list of scopes or a list of roles/clients. I created
these 2 in 1 components in order to use the same memoize function
as well as to reduce duplication of the code since a rather big
portion of the component code was going to be the same.